+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a CHANGELOG](http://keepachangelog.com/).
 
 ## [Unreleased][unreleased]
-- 
+-
+
+## [3.0.1] - 2021-08-16
+- Added support for American Express, Maestro, Mastercard, V PAY and Visa.
+- Save `CustomerIBAN` and `CustomerBIC` for Sofort payments.
 
 ## [3.0.0] - 2021-08-05
 - Updated to `pronamic/wp-pay-core`  version `3.0.0`.
@@ -97,7 +101,8 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [1.0.0] - 2015-01-19
 - First release.
 
-[unreleased]: https://github.com/wp-pay-gateways/buckaroo/compare/3.0.0...HEAD
+[unreleased]: https://github.com/wp-pay-gateways/buckaroo/compare/3.0.1...HEAD
+[3.0.1]: https://github.com/wp-pay-gateways/buckaroo/compare/3.0.0...3.0.1
 [3.0.0]: https://github.com/wp-pay-gateways/buckaroo/compare/2.2.0...3.0.0
 [2.2.0]: https://github.com/wp-pay-gateways/buckaroo/compare/2.1.2...2.2.0
 [2.1.2]: https://github.com/wp-pay-gateways/buckaroo/compare/2.1.1...2.1.2